% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ascore.R
\name{best_PTM_location_by_ascore}
\alias{best_PTM_location_by_ascore}
\title{Remap Sequence IDs in FASTA File}
\usage{
best_PTM_location_by_ascore(msnid, ascore)
}
\arguments{
\item{msnid}{(MSnID object) MS/MS ID data}

\item{ascore}{(data.frame) AScore results}
}
\value{
(MSnID object) MS/MS ID data with added AScore
}
\description{
The function remaps the IDs in the FASTA file from RefSeq to Gene symbols.
In case of multiple RefSeq sequences available, only the first longest is
retained.
}
\examples{
msnid <- add_ascore(msnid, ascore)

}
