% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_read_masic_data.R
\name{dms_read_masic_data}
\alias{dms_read_masic_data}
\title{Reading MASIC Results. PNNL system only.}
\usage{
dms_read_masic_data(DataPkgNumber = NULL, extra_metrics = FALSE)
}
\arguments{
\item{DataPkgNumber}{(Numeric or Character vector) containing Data Package ID(s) located in DMS}

\item{extra_metrics}{(logical) fetch extra metrics that MASIC extracts from dataset or not. Default is FALSE.}
}
\value{
(data.frame) with reporter ion intensities and other metrics
}
\description{
Reading MASIC output from PNNL's DMS
}
