% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_read_msgf_data.R,
%   R/read_msgf_data_from_DMS.R
\name{dms_read_msms_data}
\alias{dms_read_msms_data}
\title{Reading MSGF Results. Generic.}
\usage{
dms_read_msms_data(DataPkgNumber)

dms_read_msms_data(DataPkgNumber)
}
\arguments{
\item{DataPkgNumber}{(Numeric or Character vector) containing Data Package ID(s) located in DMS}

\item{DataPkgNumber}{(Numeric or Character vector) containing Data Package ID(s) located in DMS}
}
\value{
(MSnID) MSnID object

(MSnID) MSnID object
}
\description{
Reading MSGF output from PNNL's DMS
}
\examples{
path_to_MSGF_results <- system.file("extdata/global/msgf_output", package = "PlexedPiperTestData")
msnid <- read_msgf_data(path_to_MSGF_results)
print(msnid)
head(MSnID::psms(msnid))
msnid <- read_msgf_data(3442)
print(msnid)
head(MSnID::psms(msnid))
}
