% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_masic_data.R
\name{fetch_masic_data}
\alias{fetch_masic_data}
\title{Reading MASIC Results. PNNL system only.}
\usage{
fetch_masic_data(path_to_MASIC_results, extra_metrics = FALSE)
}
\arguments{
\item{path_to_MASIC_results}{(path string) to directory with MASIC results for all datasets}

\item{extra_metrics}{(logical) fetch extra metrics that MASIC extracts from dataset or not. Default is FALSE.}
}
\value{
(data.frame) with reporter ion intensities and other metrics
}
\description{
Reading MASIC output from PNNL's DMS
}
