% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_msgf_data_missed_cleavages.R
\name{filter_msgf_data_missed_cleavages}
\alias{filter_msgf_data_missed_cleavages}
\title{Filtering MSGF Data}
\usage{
filter_msgf_data_missed_cleavages(msnid, missed_cleavages_threshold = 2)
}
\arguments{
\item{msnid}{(MSnID object) collated MSGF output}

\item{missed_cleavages_threshold}{(numeric) Maximum acceptable number of missed cleavages. Default is 2.}
}
\value{
(MSnID object) filtered MSGF output
}
\description{
Filtering MSGF data by number of missed cleavages.
}
\examples{
path_to_MSGF_results <- system.file("extdata/global/msgf_output", package = "PlexedPiperTestData")
msnid <- read_msgf_data(path_to_MSGF_results)
msnid <- MSnID::correct_peak_selection(msnid)
show(msnid)
msnid <-  filter_msgf_data_missed_cleavages(msnid, 2) # at most 2 missed cleavages
show(msnid)
}
