% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_msgf_data.R
\name{filter_msgf_data_peptide_level}
\alias{filter_msgf_data_peptide_level}
\title{Filtering MSGF Data}
\usage{
filter_msgf_data_peptide_level(msnid, peptide_FDR_threshold = 0.01,
  n.iter.grid = 500, n.iter.nm = 100)
}
\arguments{
\item{msnid}{(MSnID object) collated MSGF output}

\item{peptide_FDR_threshold}{(numeric) Maximum acceptable FDR rate. Default is 0.01.}

\item{n.iter.grid}{(numeric) number of grid-distributed evaluation points. Default 500.}

\item{n.iter.nm}{(numeric) number of iterations for Nelder-Mead optimization algorithm. Default 100.}
}
\value{
(MSnID object) filtered MSGF output
}
\description{
Filtering MSGF data. In this implementation the function optmized both ppm and
PepQValue thresholds to achieve maximum number of peptide identifications within
given FDR constrain.
}
\examples{
path_to_MSGF_results <- system.file("extdata/global/msgf_output", package = "PlexedPiperTestData")
msnid <- read_msgf_data(path_to_MSGF_results)
msnid <- MSnID::correct_peak_selection(msnid)
show(msnid)
msnid <- filter_msgf_data_peptide_level(msnid, 0.01) # 1\% FDR
show(msnid)
}
