% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PNNL_DMS_utils.R
\name{pnnl_dms_utils}
\alias{pnnl_dms_utils}
\alias{is_PNNL_DMS_connection_successful}
\alias{get_dms_job_records}
\alias{get_tool_output_files_for_job_number}
\alias{get_output_folder_for_job_and_tool}
\alias{get_AScore_results}
\alias{get_job_records_by_dataset_package}
\alias{get_results_for_multiple_jobs}
\alias{get_results_for_multiple_jobs.dt}
\alias{get_results_for_single_job}
\alias{get_results_for_single_job.dt}
\alias{path_to_FASTA_used_by_DMS}
\title{Utilities for interacting with PNNL DMS}
\usage{
is_PNNL_DMS_connection_successful()

get_dms_job_records(jobs = NULL, datasetPttrn = "",
  experimentPttrn = "", toolPttrn = "", parPttrn = "",
  settingsPttrn = "", fastaPttrn = "", proteinOptionsPttrn = "",
  intrumentPttrn = "")

get_tool_output_files_for_job_number(jobNumber, toolName, filePattern,
  mostRecent = TRUE)

get_output_folder_for_job_and_tool(jobNumber, toolName,
  mostRecent = TRUE)

get_AScore_results(dataPkgNumber)

get_job_records_by_dataset_package(dataPkgNumber)

get_results_for_multiple_jobs(jobRecords)

get_results_for_multiple_jobs.dt(jobRecords)

get_results_for_single_job(pathToFile, fileNamePttrn)

get_results_for_single_job.dt(pathToFile, fileNamePttrn)

path_to_FASTA_used_by_DMS(data_package_number)
}
\arguments{
\item{jobs}{(integer) DMS job ID}

\item{mostRecent}{(logical) only most recent or all output files}

\item{dataPkgNumber}{(interger) data package id}

\item{mtdbName}{(string) name of mass tag database}

\item{xPttrn}{(string) substrings for names of dataset, experiment, tool, parameter file, settings file, fasta file, protein options, instrument name}
}
\description{
Fetches results of different data analysis tools.
Not useful outside of PNNL unless connected through VPN.
Works on windows out of the box. To make it work on Mac/Linux FreeTDS needs
to be installed. Mac/Linux functionality is not well tested.

\itemize{
\item \code{get_dms_job_records()}: returns job records given variety keyword patterns
\item \code{get_tool_output_files_for_job_number()}: returns the output of the tool given tool name and file pattern
\item \code{get_output_folder_for_job_and_tool()}: returns the path given job id and tool name
\item \code{get_AScore_results()}: returns Ascore results given data packge number
\item \code{get_job_records_by_dataset_package()}: returns job records given data packge number
\item \code{get_results_for_multiple_jobs()}: returns concatenated results given job numbers
\item \code{get_results_for_multiple_jobs.dt()}: returns results as concatenated data.table given job numbers
\item \code{get_results_for_single_job()}: returns results given job number
\item \code{get_results_for_single_job.dt()}: returns results as data.table given job number
}
}
\examples{
get_output_folder_for_job_and_tool(863951, "DTA_Refinery")
}
