% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AScore_results.R
\name{read_AScore_results}
\alias{read_AScore_results}
\title{Read AScore results from folder}
\usage{
read_AScore_results(path_to_AScore_results)
}
\arguments{
\item{path_to_AScore_results}{(character) Path to AScore results}
}
\value{
ascore (data.frame) AScore results
}
\description{
This function read AScore results from a local folder.
}
\examples{
ascore <- read_AScore_results("./data/ascore_outputs")
msnid <- best_PTM_location_by_ascore(msnid, ascore)

}
