% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_masic_data.R
\name{read_masic_data}
\alias{read_masic_data}
\title{Reading MASIC Results. Generic.}
\usage{
read_masic_data(path_to_MASIC_results, extra_metrics = FALSE)
}
\arguments{
\item{path_to_MASIC_results}{(path string) to directory with MASIC results for all datasets}

\item{extra_metrics}{(logical) fetch extra metrics that MASIC extracts from dataset or not. Default is FALSE.}
}
\value{
(data.frame) with reporter ion intensities and other metrics
}
\description{
Reading MASIC output from a single directory
}
\examples{
path_to_MASIC_results <- system.file("extdata/global/masic_output", package = "PlexedPiperTestData")
x <- read_masic_data(path_to_MASIC_results, extra_metrics=TRUE)
head(x)

}
