% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msgf_data_from_DMS.R
\name{read_msms_data_from_DMS}
\alias{read_msms_data_from_DMS}
\title{Reading MSGF output from PNNL's DMS}
\usage{
read_msms_data_from_DMS(DataPkgNumber)
}
\arguments{
\item{DataPkgNumber}{(Numeric or Character vector) containing Data Package ID(s) located in DMS}
}
\value{
(MSnID) MSnID object
}
\description{
Reading MSGF output from PNNL's DMS
}
\examples{
msnid <- read_msgf_data_from_DMS(3442)
print(msnid)
head(MSnID::psms(msnid))
}
