% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reporter_converter}
\alias{reporter_converter}
\title{Converter for TMT Reporter Channels}
\format{A data frame with 10 rows and 2 variables:
\describe{
  \item{ReporterName}{human-readable notation}
  \item{ReporterIon}{reporter channel names extracted by MASIC}
  ...
}}
\source{
\url{https://www.thermofisher.com/us/en/home/life-science/protein-biology/protein-mass-spectrometry-analysis/protein-quantitation-mass-spectrometry/tandem-mass-tag-systems.html}
}
\usage{
reporter_converter
}
\description{
Converter from human-readable notation of TMT reporter channels to
m/z-type (reported by MASIC).
}
\keyword{datasets}
