% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reporter_converter_new_masses}
\alias{reporter_converter_new_masses}
\title{Converter for new TMT masses}
\format{
A list with names containing new reported masses and values containing old 
   reported massess
}
\source{

}
\usage{
reporter_converter_new_masses
}
\description{
If any reporter ions have new reported masses, this list will convert them to the old
   masses.
}
\keyword{datasets}
