% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_PTM_location_by_ascore.R
\name{best_PTM_location_by_ascore}
\alias{best_PTM_location_by_ascore}
\title{Improve Phosphosite Localization}
\usage{
best_PTM_location_by_ascore(msnid, ascore)
}
\arguments{
\item{msnid}{(MSnID object) MS/MS ID data}

\item{ascore}{(data.frame) AScore results}
}
\value{
(MSnID object) MS/MS ID data with added AScore
}
\description{
Use AScore results to improve phosphosite localization.
}
\examples{
\dontrun{
# Get AScore results
path_to_ascore <- system.file("extdata/phospho/ascore_output",
                              package = "PlexedPiperTestData")
ascore <- read_AScore_results(path_to_ascore)

# MS/MS IDs
path_to_MSGF_results <- system.file("extdata/global/msgf_output",
                                    package = "PlexedPiperTestData")
msnid <- read_msgf_data(path_to_MSGF_results)

# Improve phosphosite localization
msnid <- best_PTM_location_by_ascore(msnid, ascore)
}
}
