% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_masic_data.R
\name{filter_masic_data}
\alias{filter_masic_data}
\title{Filtering MASIC Data}
\usage{
filter_masic_data(x, interference_score_threshold = 0.9, s2n_threshold = 4)
}
\arguments{
\item{x}{(data.frame) collated MASIC output}

\item{interference_score_threshold}{(numeric) The interference score reflects
the proportion of the ion population isolated for fragmentation that is due
to the targeted ion. \code{1 - InferferenceScore} is due to co-isolated species.
The higher the interference score, the cleaner the parent ion at the MS1
level. Default is 0.9.}

\item{s2n_threshold}{(numeric) S/N calculated by vendor and extracted MASIC
from raw files. Default is 4.}
}
\value{
(data.frame) filtered MASIC output
}
\description{
Filtering MASIC data by interference score and signal-to-noise ratio (S/N) of
individual channels. Note: function in its current form also drops 40 columns
out of the \code{data.frame}. E.g. for TMT10 the table contains 62 columns. After
filtering, the remaining columns are the 10 TMT channels, \code{Dataset}, and
\code{ScanNumber}.
}
\details{
Default values \code{interference_score_threshold = 0.9} and \code{s2n_threshold = 4}
are fairly stringent. We recommend at least setting
\code{interference_score_threshold = 0.5} and \code{s2n_threshold = 0}. If filtering is
not intended, use \code{filter_masic_data(x,0,0)}.
}
\examples{
\dontrun{
path_to_MASIC_results <- system.file("extdata/global/masic_output",
                                     package = "PlexedPiperTestData")
x <- read_masic_data(path_to_MASIC_results, interference_score = TRUE)
dim(x)
x1 <- filter_masic_data(x,0,0)
dim(x1)
x2 <- filter_masic_data(x)
dim(x2)
}
}
