% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_msgf_data.R
\name{filter_msgf_data}
\alias{filter_msgf_data}
\alias{filter_msgf_data_peptide_level}
\alias{filter_msgf_data_protein_level}
\title{Filtering MSGF Data}
\usage{
filter_msgf_data(
  msnid,
  level,
  fdr.max = 0.01,
  n.iter.grid = 500,
  n.iter.nm = 100
)

filter_msgf_data_peptide_level(msnid, ...)

filter_msgf_data_protein_level(msnid, ...)
}
\arguments{
\item{msnid}{(MSnID object) collated MSGF output}

\item{level}{(character) Level at which to perform FDR filter. The name of a
column in \code{psms(msnid)}. Currently, only \code{"peptide"} or \code{"accession"} are
supported.}

\item{fdr.max}{(numeric) Maximum acceptable FDR. Default is 0.01 (1\%).}

\item{n.iter.grid}{(numeric) number of grid-distributed evaluation points.}

\item{n.iter.nm}{(numeric) number of iterations for Nelder-Mead optimization
algorithm.}

\item{...}{arguments passed to \code{filter_msgf_data}.}
}
\value{
(MSnID object) filtered MSGF output
}
\description{
Filtering MSGF data. In this implementation, the peptide-level filter
optimizes both ppm and PepQValue thresholds to achieve maximum number of
peptide identifications within a given FDR constraint. The accession-level
filter optimizes based on \code{peptides_per_1000aa}, so
\code{\link{compute_num_peptides_per_1000aa}} must be used first.
}
\examples{
\dontrun{
path_to_MSGF_results <- system.file("extdata/global/msgf_output",
                                    package = "PlexedPiperTestData")
msnid <- read_msgf_data(path_to_MSGF_results)
msnid <- MSnID::correct_peak_selection(msnid)
show(msnid)
msnid <- filter_msgf_data(msnid, "peptide", 0.01) # 1\% FDR at peptide level
show(msnid)
path_to_FASTA <- system.file(
  "extdata/Rattus_norvegicus_NCBI_RefSeq_2018-04-10.fasta.gz",
  package = "PlexedPiperTestData"
)
msnid <- compute_num_peptides_per_1000aa(msnid, path_to_FASTA)
msnid <- filter_msgf_data(msnid, "accession", 0.01) # 1\% FDR at protein level
show(msnid)
}
}
\seealso{
\code{\link[MSnID]{MSnIDFilter}}
\code{\link[MSnID]{optimize_filter}}
\code{\link[MSnID]{apply_filter}}
}
