% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_id_to_quant_scan_link_table.R
\name{make_id_to_quant_scan_link_table}
\alias{make_id_to_quant_scan_link_table}
\title{Finding Links from PNNL's DMS}
\usage{
make_id_to_quant_scan_link_table(
  scanStats,
  idScanPattern = "ms2",
  quantScanPattern = "ms3",
  threshold = NULL
)
}
\arguments{
\item{pathToFile}{}
}
\value{
(data.table) with reporter ion intensities and other metrics
}
\description{
Finding Links from PNNL's DMS
}
