% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AScore_results.R
\name{read_AScore_results}
\alias{read_AScore_results}
\title{Read AScore results from folder}
\usage{
read_AScore_results(path_to_AScore_results)
}
\arguments{
\item{path_to_AScore_results}{(character) path to folder containing
a file ending in "_syn_ascore.txt".}
}
\value{
\code{\link[base]{data.frame}} of AScore results
}
\description{
This function reads AScore results from a local folder.
}
\examples{
\dontrun{
# Get AScore results
path_to_AScore_results <- system.file("extdata/phospho/ascore_output",
                                      package = "PlexedPiperTestData")
ascore <- read_AScore_results(path_to_AScore_results)
}
}
