% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_masic_data.R
\name{read_masic_data}
\alias{read_masic_data}
\title{Reading MASIC Results. Generic.}
\usage{
read_masic_data(path_to_MASIC_results, interference_score = FALSE)
}
\arguments{
\item{path_to_MASIC_results}{(character) path to directory with MASIC results
for all datasets.}

\item{interference_score}{(logical) whether to read interference score.}
}
\value{
(data.frame) with reporter ion intensities and other metrics.
}
\description{
Reading MASIC output from a single directory
}
\examples{
\dontrun{
path_to_MASIC_results <- system.file("extdata/global/masic_output",
                                     package = "PlexedPiperTestData")
x <- read_masic_data(path_to_MASIC_results, interference_score=TRUE)
head(x)
}

}
