% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_study_design.R
\name{read_study_design}
\alias{read_study_design}
\title{Utilities for reading study design files.}
\usage{
read_study_design(path_to_study_design)
}
\arguments{
\item{path_to_folder}{(string) path to folder containing study design files}
}
\description{
Fetches study design results from either local folder of DMS.
Checks that study design files are internally consistent.
DMS functionality not useful outside of PNNL unless connected through VPN.

* `read_study_design()`: returns a list of study design tables, accessible by $
}
