% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_study_design.R
\name{read_study_design}
\alias{read_study_design}
\title{Utilities for reading study design files.}
\usage{
read_study_design(path_to_study_design, prefix = character(0))
}
\arguments{
\item{path_to_study_design}{(character) path to folder containing study
design files.}

\item{prefix}{(character) optional prefix for study design tables.}
}
\value{
A list of study design tables, accessible by \code{$} or \verb{[]}.
}
\description{
Fetches study design results from local folder. Checks that
study design files are internally consistent.
}
\examples{
path_to_study_design <- system.file("extdata/study_design",
                                    package = "PlexedPiperTestData")
study_design <- read_study_design(path_to_study_design)

fractions  <- study_design$fractions
samples    <- study_design$samples
references <- study_design$references

}
