% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plexedpiper.R
\name{run_plexedpiper}
\alias{run_plexedpiper}
\title{Wrapper for PlexedPiper}
\usage{
run_plexedpiper(
  msgf_output_folder,
  fasta_file,
  masic_output_folder,
  ascore_output_folder = NULL,
  proteomics,
  study_design_folder,
  species,
  annotation,
  file_prefix = NULL,
  global_results = NULL,
  write_results_to_file = TRUE,
  output_folder = NULL,
  save_env = FALSE,
  return_results = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{msgf_output_folder}{(character) Path to MSGF+ results folder}

\item{fasta_file}{(character) Path and FASTA file name}

\item{masic_output_folder}{(data.frame) MASIC results folder}

\item{ascore_output_folder}{(data.frame) AScore results folder}

\item{proteomics}{(character) Either "pr" - proteomics, "ph" -
phosphoproteomics, "ub" - ubiquitinomics, or "ac" - acetylomics}

\item{study_design_folder}{(character) Folder containing the three study
design files: fractions.txt, samples.txt, and references.txt}

\item{species}{(character) Scientific name of species (e.g. `Rattus
norvegicus`, `Homo sapiens`, etc.)}

\item{annotation}{(character) Source for annotations: either `RefSeq` or
`Uniprot`}

\item{file_prefix}{(character) Prefix for the file name outputs}

\item{global_results}{(character) Only for PTM experiments. Ratio results
from a global protein abundance experiment. If provided, it will infer
parsimonious set of accessions.}

\item{write_results_to_file}{(logical) Whether to write the results to files.}

\item{output_folder}{(character) Output folder name to save results. If not
provided, it will save it to the current directory.}

\item{save_env}{(logical) Whether to save the R environment to the output
folder.}

\item{return_results}{(logical) Whether to return both ratio and rii results.}

\item{verbose}{(logical) Whether to show messages describing steps.}
}
\value{
(list) If `return_results` is `TRUE`, it returns list with ratio and
  RII data frames.
}
\description{
A wrapper function for PlexedPiper that performs all the steps
  necessary to go from MS-GF+ and MASIC data to Reporter Ion Intensity (RII)
  and Ratio Results tables used by the MoTrPAC Bioinformatics
  Center (BIC).
}
\examples{
\dontrun{
results <- run_plexedpiper(msgf_output_folder = "~/path/to/msgfplus/,
                           fasta_file  = "~/path/to/fasta/sequence.fasta,
                           masic_output_folder = "~/path/to/masic-results/",
                           ascore_output_folder = "~/path/to/ascore-results/",
                           proteomics = "ph",
                           study_design_folder = "~/path/to/study-design/",
                           species = "Rattus norvegicus",
                           annotation = "RefSeq",
                           global_results = "~/path/to/global/ratio.txt",
                           output_folder = "~/path/to/pp-results/",
                           file_prefix = "msgfplus-pp-results",
                           return_results = TRUE,
                           verbose = TRUE)
}
}
