% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plexedpiper.R
\name{run_plexedpiper}
\alias{run_plexedpiper}
\title{Run PlexedPiper}
\usage{
run_plexedpiper(
  msgf_output_folder,
  fasta_file,
  masic_output_folder,
  ascore_output_folder = NULL,
  proteomics,
  study_design_folder,
  species,
  annotation,
  file_prefix = NULL,
  global_results = NULL,
  write_results_to_file = TRUE,
  output_folder = NULL,
  save_env = FALSE,
  return_results = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{msgf_output_folder}{(MSnID) Path to MSGF+ results folder}

\item{fasta_file}{(character) Path and fasta file name}

\item{masic_output_folder}{(data.frame) Masic results folder}

\item{ascore_output_folder}{(data.frame) AScore result folder}

\item{proteomics}{(character) One of the following:}

\item{study_design_folder}{(list) Study design folder}

\item{species}{(character) Scientific name of a specie (e.g. `Rattus norvegicus`, `Homo sapiens`)}

\item{annotation}{(character) Source for annotations: either `RefSeq` or `Uniprot`}

\item{file_prefix}{(character) Prefix for the file name outputs}

\item{global_results}{(character) Only for PTM experiments. 
Ratio results from a global protein abundance experiment. 
If provided, it will infer parsimonious set of accessions. 
Default is `NULL`}

\item{write_results_to_file}{(logical) if `TRUE`, write the results to files}

\item{output_folder}{(character) Output folder name to save results. 
(if not provided it will save it to the current directory)}

\item{save_env}{(logical) If `TRUE` it saves the R environment to the output folder}

\item{return_results}{(logical) If `TRUE` return both ratio and rii results (default `FALSE`)}

\item{verbose}{(logical) `TRUE` (default) shows messages}
}
\value{
(list) If `return_results` is TRUE, it returns list with ration and rii data frames
}
\description{
Provide outputs from MSGF+, MASIC, ASCORE (if PTM), 
plus other parameters including fasta sequence db, species, annotation, 
and it returns both the Reported Ion Intensity (RII) and ratio results
}
\examples{
\dontrun{
results <- run_pp(msgf_output_folder = msgf_output_folder,
                  fasta_file  = fasta_file,
                  masic_output_folder = masic_output_folder,
                  ascore_output_folder = ascore_output_folder,
                  proteomics = tolower(proteomics),
                  study_design_folder = study_design_folder,
                  species = species,
                  annotation = annotation,
                  global_results = plexedpiper_global_results_ratio,
                  output_folder = plexedpiper_output_folder,
                  file_prefix = plexedpiper_output_name_prefix,
                  return_results = TRUE,
                  verbose = TRUE)
}
}
