% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plexedpiper.R
\name{run_plexedpiper}
\alias{run_plexedpiper}
\title{Wrapper for PlexedPiper}
\usage{
run_plexedpiper(
  msgf_output_folder,
  fasta_file,
  masic_output_folder,
  ascore_output_folder = NULL,
  proteomics,
  study_design_folder,
  prefix = character(0),
  species,
  annotation,
  file_prefix = NULL,
  unique_only = FALSE,
  global_results = NULL,
  refine_prior = FALSE,
  write_results_to_file = TRUE,
  output_folder = NULL,
  save_env = FALSE,
  return_results = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{msgf_output_folder}{(character) Path to MS-GF+ results folder(s).}

\item{fasta_file}{(character) Path to FASTA file. If multiple MS-GF+ results
folders are provided, all should be searched against the same
protein database.}

\item{masic_output_folder}{(character) MASIC results folder(s).}

\item{ascore_output_folder}{(character) AScore results folder(s).}

\item{proteomics}{(character) Either "pr" - proteomics, "ph" -
phosphoproteomics, "ub" - ubiquitinomics, or "ac" - acetylomics.}

\item{study_design_folder}{(character) Folder containing the three study
design tables: fractions.txt, samples.txt, and references.txt.}

\item{prefix}{(character) optional prefix for study design tables.}

\item{species}{(character) Scientific name of species (e.g. "Rattus
norvegicus", "Homo sapiens", etc.).}

\item{annotation}{(character) Source for annotations: either \code{RefSeq},
\code{UniProt}, or \code{Gencode} (case insensitive).}

\item{file_prefix}{(character) Prefix for the file name outputs.}

\item{unique_only}{(logical) Whether to discard peptides that match multiple
proteins in the parsimonious protein inference step. Default \code{FALSE}.}

\item{global_results}{(character) Only for PTM experiments. Ratio results
from a global protein abundance experiment. If provided, prioritized
inference will be performed. Otherwise, parsimonious inference is performed
without a prior. See \code{\link[MSnID]{infer_parsimonious_accessions}} for
details.}

\item{refine_prior}{(logical) if \code{FALSE} (default), peptides are allowed
to match multiple proteins in the prior. That is, the greedy set cover
algorithm is only applied to the set of proteins not in the prior. If
\code{TRUE}, the algorithm is applied to the prior and non-prior sets
separately before combining. See
\code{\link[MSnID]{infer_parsimonious_accessions}} for more details.}

\item{write_results_to_file}{(logical) Whether to write the results to files.}

\item{output_folder}{(character) Output folder name to save results. If not
provided, it will save it to the current directory.}

\item{save_env}{(logical) Whether to save the R environment to the output
folder.}

\item{return_results}{(logical) Whether to return the ratio and rii results.}

\item{verbose}{(logical) Whether to show messages describing steps.}
}
\value{
(list) If \code{return_results} is \code{TRUE}, it returns list with ratio and
RII data frames.
}
\description{
A wrapper function for PlexedPiper that performs all the steps
necessary to go from MS-GF+ and MASIC data to Reporter Ion Intensity (RII)
and Ratio Results tables used by the MoTrPAC Bioinformatics Center (BIC).
}
\examples{
\dontrun{
# Example with pseudo-paths
results <- run_plexedpiper(msgf_output_folder = "~/path/to/msgfplus/",
                           fasta_file  = "~/path/to/fasta/sequence.fasta",
                           masic_output_folder = "~/path/to/masic-results/",
                           ascore_output_folder = "~/path/to/ascore-results/",
                           proteomics = "ph",
                           study_design_folder = "~/path/to/study-design/",
                           species = "Rattus norvegicus",
                           annotation = "RefSeq",
                           global_results = "~/path/to/global/ratio.txt",
                           unique_only = FALSE,
                           refine_prior = FALSE,
                           output_folder = "~/path/to/pp-results/",
                           file_prefix = "msgfplus-pp-results",
                           return_results = TRUE,
                           verbose = TRUE)
}
}
